added Auth parent class
Added a new abstract class called Auth from which methods shared by credentials and token can be managed.
